move to cc.strings
not quite working, one failing test with an edge case on word wrapping
hopefully someone better at strings than me can help
